Add def special form definition
