--- core/core.go
@@ -1,3 +1,0 @@
-// Package core provides core functionalities built into Sabre system including
-// macro system, built-in special forms etc.
-package core
